pyproject.toml
[project]
name = "soop-discord-bot"
version = "0.1.0"
description = "Community Discord bot that posts SOOP live notifications."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "Community" }]

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "py-cord>=2.5.0",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "ruff>=0.4",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
